% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjplot2pdf.R
\name{build_pdf}
\alias{build_pdf}
\title{Build pdf to check}
\usage{
build_pdf(file, silent, pdf_file, pdf_file_output, filename)
}
\arguments{
\item{file}{TeX file}
}
\value{
pdf file
}
\description{
Build pdf to check
}
\keyword{internal}
