% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjplot2pdf.R
\name{clean_debris}
\alias{clean_debris}
\title{Clean up latex debris}
\usage{
clean_debris(filename)
}
\arguments{
\item{filename}{file path to clean up}
}
\value{
pdf file
}
\description{
Clean up latex debris
}
\keyword{internal}
