% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2pdf.R, R/tex2Rmd.R
\name{html2pdf}
\alias{html2pdf}
\alias{tex2Rmd}
\title{html2pdf}
\usage{
html2pdf(
  filename,
  table_width = 30,
  page_width = 8.5,
  page_height = 11,
  build_pdf = FALSE,
  clean = FALSE,
  style = TRUE,
  silent = TRUE,
  OS = NULL,
  path_w2l = NULL,
  name_table = "table1"
)

tex2Rmd(filename, output_file = "table.txt")
}
\arguments{
\item{filename}{TeX file to read}

\item{table_width}{Width of the table}

\item{page_width}{In inches}

\item{page_height}{In inches}

\item{build_pdf}{TRUE/FALSE}

\item{clean}{TRUE/FALSE clean intermediate files}

\item{style}{makes table somewhat prettier}

\item{silent}{TRUE/FALSE}

\item{OS}{String identifying operating system (Linux, macOS or Windows)}

\item{path_w2l}{Path to the Writer2latex files}
}
\value{
TeX file

txt file
}
\description{
Convert \code{\link[sjPlot:tab_model]{sjPlot::tab_model()}} html tables to tex and pdf

extracts the table code
(from \verb{\\begin\{longtable\}} to \verb{\\end\{longtable\}})
from the Tex file which is
generated with \link{html2pdf}
}
\examples{
\dontrun{
model = lmer(mpg ~ cyl * disp + (1|vs), mtcars)

# We save the sjPlot table to an .html file
sjPlot::tab_model(
  model,
  show.r2 = TRUE,
  show.icc = FALSE,
  show.re.var = FALSE,
  p.style = "scientific",
  emph.p = TRUE,
  file = "temp.html")

# Create tex and pdf
html2pdf(filename = "temp.html", table_width = 13, silent = TRUE,
          style = TRUE, build_pdf = TRUE, clean = TRUE)
}
}
