% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjplot2pdf.R
\name{sjplot2pdf}
\alias{sjplot2pdf}
\title{sjplot2pdf}
\usage{
sjplot2pdf(
  filename,
  table_width = 30,
  page_width = 8.5,
  page_height = 11,
  build_pdf = FALSE,
  clean = FALSE,
  style = TRUE,
  silent = TRUE,
  mac = TRUE,
  path_w2l = "."
)
}
\arguments{
\item{filename}{Name of the sjPlot table object}

\item{table_width}{With of the table}

\item{page_width}{In inches}

\item{page_height}{In inches}

\item{build_pdf}{TRUE/FALSE}

\item{clean}{TRUE/FALSE}

\item{style}{makes table somewhat prettier}

\item{silent}{TRUE/FALSE}

\item{mac}{when run on a Mac}

\item{path_w2l}{Path to the Writer2latex files}
}
\value{
TeX file
}
\description{
Convert \code{\link[sjPlot:tab_model]{sjPlot::tab_model()}} html tables to tex and pdf
}
\examples{
\dontrun{
model = lmer(mpg ~ cyl * disp + (1|vs), mtcars)

# We save the sjPlot table to an .html file
sjPlot::tab_model(
  model,
  show.r2 = TRUE,
  show.icc = FALSE,
  show.re.var = FALSE,
  p.style = "scientific",
  emph.p = TRUE,
  file = "temp.html")

# Create tex and pdf
html2pdf(filename = "temp.html", table_width = 13, silent = TRUE,
          style = TRUE, build_pdf = TRUE, clean = TRUE)
}
}
