% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2pdf.R
\name{clean_tex}
\alias{clean_tex}
\title{Minimal improvements to format}
\usage{
clean_tex(file, table_width, page_width, page_height, name_table = "table1")
}
\arguments{
\item{file}{TeX file}

\item{table_width}{With of the table}

\item{page_width}{In inches}

\item{page_height}{In inches}
}
\value{
TeX file
}
\description{
Minimal improvements to format
}
\keyword{internal}
